Correection

+++ 0x0B-malloc_free/100-argstostr.c
@@ -15,7 +15,7 @@ char *argstostr(int ac, char **av)
 
 	if (ac == 0 || av == NULL)
 		return (NULL);
-	while (i < ac)
+	while (i < (size_t) ac)
 	{
 		len += (strlen(av[i]) + 1);
 		i++;
